requirements.txt
streamlit
pandas
altair
openai>=1.50.0
Pillow